requirements.txt
zendriver
openai
tenacity
wxauto
python-dotenv
pytest
pytest-cov